+++ CHANGELOG.rst
@@ -4,6 +4,39 @@ Changelog
 
 .. _newest-changelog:
 
+0.9.0a2
+========
+
+Major Features and Improvements
+-------------------------------
+- Save results by pickling, unpickling a frozen (`FitResult.freeze()`) result and using
+  `zfit.param.set_values(params, result)` to set the values of `params`.
+
+
+
+Deprecations
+-------------
+- the default name of the uncertainty methods `hesse` and `errors` depended on
+  the method used (such as 'minuit_hesse', 'zfit_errors' etc.) and would be the exact method name.
+  New names are now 'hesse' and 'errors', independent of the method used. This reflects better that the
+  methods, while internally different, produce the same result.
+  To update, use 'hesse' instead of 'minuit_hesse' or 'hesse_np' and 'errors' instead of 'zfit_errors'
+  or 'minuit_minos' in order to access the uncertainties in the fitresult.
+  Currently, the old names are still available for backwards compatibility.
+  If a name was explicitly chosen in the error method, nothing changed.
+
+Bug fixes and small changes
+---------------------------
+- KDE datasets are now correctly mirrored around observable space limits
+- multinomial sampling would return wrong results when invoked multiple times in graph mode due to
+  a non-dynamic shape. This is fixed and the sampling is now working as expected.
+- increase precision in FitResult string representation and add that the value is rounded
+
+
+Thanks
+------
+ - schmitse for finding and fixing a mirroring bug in the KDEs
+ - Sebastian Bysiak for finding a bug in the multinomial sampling
 
 0.9.0a0
 ========
@@ -15,6 +48,8 @@ Major Features and Improvements
   binned losses. TODO: extend to include changes/point to binned introduction.
 - new Poisson PDF
 - added Poisson constraint, LogNormal Constraint
+- Save results by pickling, unpickling a frozen (`FitResult.freeze()`) result and using
+  `zfit.param.set_values(params, result)` to set the values of `params`.
 
 Breaking changes
 ------------------
@@ -23,7 +58,7 @@ Breaking changes
 - `norm_range` is now called `norm` and should be replaced everywhere if possible. This will break in
   the future.
 
-Depreceations
+Deprecation
 -------------
 
 Bug fixes and small changes
@@ -260,7 +295,7 @@ Breaking changes
 - Creating a ``zfit.Parameter``, usign ``set_value`` or ``set_values`` now raises a ``ValueError``
   if the value is outside the limits. Use ``assign`` to suppress it.
 
-Depreceations
+Deprecation
 -------------
 - strategy to minimizer should now be a class, not an instance anymore.
 
@@ -324,7 +359,7 @@ Major Features and Improvements
 Breaking changes
 ------------------
 
-Depreceations
+Deprecation
 -------------
 
 Bug fixes and small changes
@@ -364,9 +399,6 @@ Breaking changes
 - Minuit uses its own, internal gradient by default. To change this back, use ``use_minuit_grad=False``
 - ``minimize(params=...)`` now filters correctly non-floating parameters.
 - ``z.log`` has been moved to ``z.math.log`` (following TF)
-
-Depreceations
--------------
 
 
 Bug fixes and small changes